AP: Don't increment auth_transaction upon SAE authentication failure
802.11-2016 12.4.7.6 specifies:

An SAE Commit message with a status code not equal to SUCCESS shall
indicate that a peer rejects a previously sent SAE Commit message.

An SAE Confirm message, with a status code not equal to SUCCESS, shall
indicate that a peer rejects a previously sent SAE Confirm message.

Thus when SAE authentication failure happens, authentication transaction
sequence number should not be incremented.

Change-Id: Ie5254741dd47147d8f7bcb684747b07a82ce650e
CRs-Fixed: 2985733