wifi(hidl): Test channel HT40 capability
This commit aims to always enable HE80 when 5GHz band
is used and acs is disabled.

When acs is disabled, framework randomly chooses a channel.
Since "[HT40+][HT40-]" is only used in acs case, so this
commit test channel ht40 capability to decide HT40+ or HT40-.

Meanwhile, for FIRST channel that larger than 157 (i.e. 165),
override it to 157.

CRs-Fixed: 3071466
Change-Id: I553bd466085041f31caa26f6914b8e3b3b2c1916