hostapd(hidl): Set dtim_period=1 for hostapd configuration
The default value for the hostapd configuration parameter
dtim_period is likely a more or less arbitrarily picked out
value that happened to be something that common APs used
15-20 years ago, i.e., there is not really any specific reason
for that default.

DTIM period should be configured in the initial hostapd
configuration to the value that is desired for the network.

Per OEM requirement, set dtim_period=1 for hostapd configuration.

CRs-Fixed: 2900054
Change-Id: I600f8cbe512d20417f1da50e5d605af59551819e