Revert "STA OBSS: Add check for overlapping BSSs"
This reverts commit 3204795d7a43bcbda8b695dd96d11f3000de7340.

The commit adds an additional check that checks for overlapping BSSs in
addition to the existing 40 MHz intolerance subfield checks. The commit
cites IEEE Std 802.11-2016, 11.16.12, which defines the proper behavior
for a 20/40 MHz HT STA and AP, but the standard actually doesn't say
anything about overlapping BSSs. Specifically, the standard states that
the only BSSs that belong in the Intolerant channel report are those
that satisfy trigger event A, defined as channels with BSSs that don't
contain the HT capabilities element (which wpa_supplicant already did
before). Note that we also include channels with BSSs that have the 40
MHz intolerance bit set in the Intolerant channel report.

CRs-Fixed: 3221448
Change-Id: Icc44022b76c4df9859d5edfa4ccf7aca6ae740f9
Signed-off-by: Guisen Yang <quic_guiseny@quicinc.com>