P2P: Track peer 6 GHz capability more robustly
It's necessary to maintain knowledge of the 6 GHz capability of the
peer. Since the Device Capability field migth change between frames
depending on the context in which they are used, loooking at the last
received message might not always provide accurate information.

Add supports_6ghz bool variable in struct p2p_device, initialize it to
false and set to true if the P2P_DEV_CAPAB_6GHZ_BAND_CAPABLE bit is set
to 1 in any P2P frame that includes the P2P Capability attribute. This
boolean would not be cleared to false at any point in time so that the
info doesn't disappear dynamically.

Change-Id: Ieff9ddabfd4b46c12f940b5d3a7a6c4540c28ea7
Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Git-repo: git://w1.fi/hostap.git
Git-commit: 9c830d917800bff3a53c157063afcf9cd3087399
CRs-Fixed: 3321335