AP: Enable H2E on 6 GHz when SAE is used
Even if the use of H2E isn't strictly mandatory when using SAE on 6 GHz,
WPA3-Personal pushes it on 6 GHz. So lets automatically enable it by
setting sae_pwe=2. This will allow both the hunting-and-pecking and
hash-to-element to work (and be backward compatible).

Git-Repo: git://w1.fi/srv/git/hostap.git
Git-commit: 20bfd4feb3935a562e0c9de73cc9c8f1facec25c
Change-Id: I923686787d243429b7cd0c09a1de70bb2100d291
Signed-off-by: Nicolas Escande <nico.escande@gmail.com>