P2P: Skip 6 GHz band directly if 6 GHz P2P is disabled
If 6 GHz is supported by the device but 6 GHz P2P is disabled, P2P
invitation would fail if the GO select an operating channel which is not
the preferred channel. The root cause is that the 5 GHz and 6 GHz bands
are both HOSTAPD_MODE_IEEE80211A so the 5 GHz channels would be added
twice for the P2P Client's following scanning frequency list. This will
cause scanning to fail with -EINVAL. Fix this by adding the 5 GHz
channels only once.

Signed-off-by: Hector Jiang <jianghaitao@zeku.com>
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: e9627f8c32f816ed28a60d96d006e4d2338cc3c8

Change-Id: If8d9067a423e44096564a6047bd7dc1fef3b94de
CRs-Fixed: 3314073