P2P: Include only 6 GHz PSCs in full scan
As P2P GOs are not expected to be collocated, i.e., they are not
expected to be announced in the RNR element of other APs, they can
operate only on preferred scanning channels (PSCs).

When performing a full scan for P2P discovery, include only the 6 GHz
PSCs (if supported) to avoid scanning channels on which P2P GOs are not
expected to reside.

While at it also fix couple of places that missed including 60 GHz
channels in P2P full scan.

Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 50a9b7d3d35593fbec4b6fa2a4464c4f309cf625

Change-Id: I24e4326c1603a8e4bb4155c676ec5c69fb8c0a40
CRs-Fixed: 3314073