wpa_supplicant: Restart the eap_proxy during daemonization
When the wpa_supplicant is started using ctrlapp_dut, it uses -B
flag in the command. And this flag daemonizes the wpa_supplicant.
The test case then tries to kill the wpa_supplicant and
reinitialize it.
Initially when the wpa_supplicant was started, eap_proxy gets
created as a child thread under this parent. And when the
wpa_supplicant gets daemonized and created as a new process, the
eap_proxy thread also gets terminated and this will cause the
pthread_join to block the wpa_supplicant deinit process.
And this results in error when wpa_supplicant tries to restart.

To fix this restart the eap_proxy thread whenever wpa_supplicant
is daemonized.

CRs-Fixed: 3390739
Change-Id: I24abfec5ed27e68db3c07d21c00b4cbd32e24170