wpa_ctrl: set proper file permissions for client socket file
The standard umask of Linux on OWRT is 0x0022, which has
been changed to 0x0 by adb daemon. So in adb shell, it's
0x0, and hostapd_cli/wpa_cli works well; but in serial
console, it's the standard value 0x0022, then the file
permissions of the client socket file which is created
by hostapd_cli is (default & ~0x0022), with no write
permission for group and others, which results in
permission denied when the server side (wpa_supplicant
or hostapd) is not running with the same user.
To fix this issue, set client socket file permissions
to make sure the server side (wpa_supplicant or hostapd)
can reply to the control interface messages.

Change-Id: Icb032dce1a69789c692005f8c5484b09379e725c
CRs-Fixed: 3434308