Fix hostapd interface cleanup with multiple interfaces
interfaces.iface[i] might be NULL when going through the cleanup of all
remaining interfaces at the end of the process termination. The changes
for clearing drv_priv in AP MLD cause that cleanup process to crash on
dereferencing a NULL pointer.

Fix this by explicitly checking that the interface context is available
before trying to clear the pointer to driver data.

Fixes: 7fa99b3246d1 ("AP: Allow starting multiple interfaces within single MLD")
Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: dd1330b502ffdb076abaef4f16764078dd2e4bf9
Signed-off-by: Zhongyue Gao <quic_zhongao@quicinc.com>
Change-Id: Ic1e1311369457aa8fd3c5402f002343717d594ed