Update BSS entry on roaming only for actual BSS frequency change
Skip the full BSS table update on association event if the event is for
a roaming case to a different BSS and the BSS entry for the target is
already available and for the current operating channel. This avoids
undesired latency when processing the association event.
Also optimize internal BSS table updates based on a specific BSSID.
When wpa_supplicant needed to update the internal BSS table with the
latest scan results from the driver, it fetched all BSSs and processed
them all. This is unnecessary for cases where an update is needed only
for a specific BSS. Optimize this by filtering out the unnecessary
entries from the results.

Change-Id: I080460a86f7ae77bace021efe63249e1f61a831f
CRs-Fixed: 3771264