RSNO: Use the RSN Selection element to indicate which variant was used
This replaces the use of the RSNE Override and RSNE Override 2 elements
with empty payload to indicate which RSNE variant was used.

In addition, this adds stricter validation of the RSNE in
(Re)Association Request frame to allow only the pairwise cipher suites
and AKMs listed in the indicated RSNE variant to be used.

Change-Id: I7181f78586e6307b02995a9c4b545d5d3d0e5108
Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>