PASN: Add support to reject PASN auth 1 based on user input
When the PASN implementation is used as a library for Wi-Fi Aware
applications there is need to reject the PASN auth 1 frame based on the
user input. Add support to send PASN auth 2 with failure status for
auth 1 frame received to address this need.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 8f21cdf9d765f2d1462fef8f96dcf2e878f938dc
Signed-off-by: Zhongyue Gao <quic_zhongao@quicinc.com>
Change-Id: I18cdc7c698b855ad5a06199c55b492318bbb5a49