AP MLD: Fix PN/IPN/BIPN for group rekeying
wpa_auth_get_seqnum() for ML group rekeying needs to be skipped in the
same way as it is done for non-ML cases to avoid indicating old values
and resulting in group frames being dropped as replays. The simple check
for gsm->wpa_group_state != WPA_GROUP_SETKEYS (as is done for non-ML) is
not sufficient for this since the per-link Authenticator states are not
strictly synchronized and the state change happens in the middle of this
step.

Fixes: 137b85509248 ("MLO: Mechanism for fetching group key information for the links")
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Git-repo: git://w1.fi/hostap.git
Git-commit: 8c9b9ccb491b053cc6407ba3baba6c442cfc7784
CRs-Fixed: 4071127
Change-Id: Ib62899d36abc0394aec40e507b6ed08d3fbb8f18