Remove STA entries if association is not completed in 60 seconds
While the IEEE 802.11 standard allows STAs to authenticate with multiple
APs and later associate with one such AP, it is not really good for an
AP to maintain STA entries for not fully associated STA for significant
amount of time. Time out such STA entries in hostapd to clean state and
resources.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
CRs-Fixed: 3861126
Git-repo: git://w1.fi/hostap.git
Git-commit: 07ee3e8b0eb4026a0affb209b3d44ba232e532e9
Change-Id: Ib9b10827114399316a38d80382f392a6548b3337