AP MLD: Cancel ap_sta_assoc_timeout for partner links
An earlier commit implements a timeout handler to remove unassociated
sta entries. For AP MLD, when a non-AP MLD is associated, the timeout
handler needs to be removed in its partner links, too. The check for
WLAN_STA_ASSOC is supposed to skip such cases in ap_sta_assoc_timeout(),
but it is better to cancel this timeout altogether from all links.

Fixes: 07ee3e8b0eb4 ("Remove STA entries if association is not completed in 60 seconds")
Signed-off-by: Chenming Huang <quic_chenhuan@quicinc.com>
Git-repo: git://w1.fi/hostap.git
Git-commit: 298592db0eed2b293555a3abf040e5f5e8815b6f

CRs-Fixed: 3932573
Change-Id: I63c3af39f37077e4c9d14983f7934c87aa9b899e