OpenSSL: Avoid use of an uninitialized array
Initialize the skip_buf[] array before using it with EVP_CipherUpdate()
to skip the initial segment of RC4 output. This does not change actual
behavior since the output of that call is not used and it is only there
for changing the internal state of the RC4 cipher. However, this avoids
uninitialized element issues reported in MISRA.

Change-Id: I7e503cc8119f23cbc2d44d4c4d7f28cecdc785e1
Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Git-repo: git://w1.fi/hostap.git
Git-commit: 9fa657a73ecfe6618c7f50d53b0afe4a25c710c9
CRs-Fixed: 4124749