P2P GO: Use driver preferred frequency instead of ACS when band is specified
ACS was enabled by default when the band is specified as 2 or 5 in the
"freq" parameter for P2P_GROUP_ADD control interface command. Use the
driver preferred frequency list instead of ACS when the driver preferred
frequency list is available. This allows the GO to be started more
quickly.

Signed-off-by: Kavita Kavita <kkavita@qti.qualcomm.com>
CRs-Fixed: 4135133
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 0306618ed2d7032a9116ea53dfbd1f02780bacfd
Change-Id: Ida128f5fd1ad4ae5f1e3df863f48580fb41fc2ec