P2P: Allow upgrade to 160 MHz if 80 MHz allowed and AUTO_BW flag is set
Upgrade to 160 MHz when AUTO_BW flag is set in contiguous frequency
range. This is needed to allow regdb rules for 160 MHz channels to be
used correctly when some of the differences between the 80 MHz parts of
the full channel has resulted in AUTO_BW flag being used.

Signed-off-by: Chenming Huang <chenhuan@oss.qualcomm.com>
Git-repo: git://w1.fi/hostap.git
Git-commit: 02ccc95abb0090510548b0f900459916e3493e7c
CRs-Fixed: 4173726
Change-Id: I7002aeabd907d7ae188f07a2eacded6d9817a768