PASN: Reuse existing PTKSA cache keys instead of redoing PASN
When a successful PASN exchange has already occurred with a peer,
the derived keys are stored in the PTKSA cache. If another PASN
request is received for the same peer, these cached keys should
be reused. However, the cache was previously bypassed due to the
incorrect pairwise cipher argument.

Ensure the correct parameters are set in advance, from the RSNE data
allowing the cache to be used when appropriate.

CRs-Fixed: 4196217
Git-repo: git://w1.fi/srv/git/hostap.git
Git-commit: 2b996453ff905b41d3e7a9363e8e2d9d75f04671
Change-Id: I483fd0248d38cca390d5c5052861983e8f115484
Signed-off-by: Peddolla Harshavardhan Reddy <peddolla@qti.qualcomm.com>