GAS: Prefer unique dialog token value across all pending queries
Previously, the dialog token for a new GAS query was selected in a
manner that would be random and unique for the specific peer. However,
with the introduction of support for AP MLDs, multiple different
addresses could have ended up using the same dialog token value for the
same peer.

Make duplicate dialog token uses for AP MLDs less likely by first trying
to find a random dialog token value that is not used in any pending GAS
query for any peer. Only if that does not succeed, fall back to unique
value for the specific peer address to maintain previous behavior.

Signed-off-by: Jouni Malinen <jouni.malinen@oss.qualcomm.com>
Git-repo: git://w1.fi/hostap.git
Git-commit: 16f1b8b72ba46415dbd4383af2685a75640e07e8
CRs-Fixed: 4229230
Change-Id: I6cc1e36ad4fd90dfbe17bff857095f8206ff625f