GAS: Do not select pending query that has not yet been sent
An unexpected GAS response for a query that was scheduled to be sent,
but not yet actually transmitted, could result in unexpected behavior.
This could end up selecting the pending query for which the radio work
had not yet been started when processing the received frame and that
could result in freeing the pending query but not being able to
terminate the radio work that still has a reference to the pending
query.

Do not allow a pending GAS query to match a received frame if that query
has not yet been transmitted to avoid this corner case.

Signed-off-by: Jouni Malinen <jouni.malinen@oss.qualcomm.com>
Git-repo: git://w1.fi/hostap.git
Git-commit: ebb0cf919c466463640d17c00ad0ee2caae17e13
CRs-Fixed: 4229230
Change-Id: If6dd96458602a6c35b68b53cf24cd0f9337a969c