PASN: Fix buffer tailroom validation in wpa_pasn_add_extra_ies()
The length of the additional elements was not used correctly, so the
check for remaining tailroom would not have caught cases where there is
not enough remaining room in the buffer and the following
wpabuf_put_data() operation would have resulted in terminating the
process due to the additional check to prevent buffer overflows.

The existing use cases for this function within wpa_supplicant do not
seem to generate buffers that would be even close to reaching this limit
due to large buffer size used for the target. Anyway, this check needs
to be fixed to avoid any potential issues in the future or in external
uses for this function.

Fixes: b1ed44b6a699 ("PASN: Allow extra elements to be added into PASN Authentication frames")
Signed-off-by: Jouni Malinen <jouni.malinen@oss.qualcomm.com>

CRs-Fixed: 4319406
Git-repo: git://w1.fi/hostap.git
Git-commit: 1cf000222f337929559423746502161a8cf16c5d
Change-Id: Icdb69cbe058566d5749552446f7ccc5d560909e2