PASN: Fix buffer tailroom validation in wpa_pasn_add_wrapped_data()
While the initial tailroom checks covers the unfragmented case
accurately, the length of the fragment header was not counted correctly
for the case where the Wrapped Data element needs to be fragmented. This
could theoretically result in missing a case where the target buffer is
a bit shorter than all the needed fragments and the following
wpabuf_put*() operation could resulted in terminating the process due to
the additional check to prevent buffer overflows.

The existing use cases for this function within wpa_supplicant do not
seem to generate buffers that would be even close to reaching this limit
due to large buffer size used for the target. Anyway, this check needs
to be fixed to avoid any potential issues in the future or in external
uses for this function.

Fixes: 9ce123cdbf82 ("PASN: Add common Authentication frame build/validation functions")
Signed-off-by: Jouni Malinen <jouni.malinen@oss.qualcomm.com>

CRs-Fixed: 4363000
Git-repo: git://w1.fi/hostap.git
Git-commit: bb41fdcf2f8e2fb9677c3424036860afc146c297
Change-Id: I835ab46dd573b58e82b02fe925a41c719158b1e8