PASN: Support specifying variable length KCK
The current PASN definition uses a fixed length KCK, but that does not
feel ideal for some use cases, e.g., with the PASN-based EPPKE. Add
support for specifying the KCK length when calculating the MIC for PASN
Authentication frames. The actual calculation of the MIC is still
requiring the hardcoded 32 octet length to be used, i.e., this does not
change the actual behavior yet, but is a step towards making it easier
to support variable length KCK in the future, if needed.

Signed-off-by: Sai Pratyusha Magam <smagam@qti.qualcomm.com>
Signed-off-by: Ainy Kumari <ainy.kumari@oss.qualcomm.com>
CRs-Fixed: 4390685
Git-repo: git://w1.fi/hostap.git
Git-commit: c4776072a93913730580ef6a74282856012bf067
Change-Id: I43fbb226cbf862b975707c0ee3d4d4e1878d49dd