Fix incorrect calculation of max supported link speed
The maximum supported Tx link speed reported by the system is incorrect
due to a logical error in the negotiation of the maximum Number of
Spatial Streams (NSS). The calculation was incorrectly comparing the
STA's and AP's capabilities directly (e.g., Tx-to-Tx).

The correct logic must compare the transmitter of one device with the
receiver of the other. This patch fixes the calculation as follows:
- Max Rx NSS (STA) is the minimum of the STA's Rx and the AP's Tx
capabilities.
- Max Tx NSS (STA) is the minimum of the STA's Tx and the AP's Rx
capabilities.

This ensures the correct maximum supported link speed is calculated and
reported.

Bug: 465366239
Test: Verified by QCA
Change-Id: Ib3388f79381f94309740d64bb8f07ef376c245ff
Signed-off-by: Henry Yen <henryyen@google.com>
CRs-Fixed: 4364111
Git-repo: git://w1.fi/hostap.git
Git-commit: c9417ec62a2d776408d471f14d39f16606f52b57